Move logger and report files into folders for seperation
