+++ .gitignore
@@ -1,6 +1,6 @@
 node_modules
 dist
-reports
+/reports
 npm-debug.log
 .idea
 test/*.xml
