+++ src/reports/index.js
@@ -2,38 +2,14 @@
 import _ from 'underscore';
 import fs from 'fs';
 import mkdirp from 'mkdirp';
-import logger from './logger.js';
+import logger from '../logger';
+
+import GenerateReport from './generateReport';
 
 const defaultOptions = {
   fileName: 'report',
   reportType: 'json',
   location: ''
-};
-
-const generateReport = (reports, seperator) => {
-
-  let output = '';
-
-  _.each(reports, (report, fileName) => {
-    const headings = ['heading', 'issue', 'element', 'id', 'class', 'line', 'column', 'description'].join(seperator);
-
-    output += `${fileName}\n`;
-    output += `${headings}\n`;
-
-    report.messageLog.forEach((message) => {
-      output += message.heading + seperator;
-      output += message.issue + seperator;
-      output += message.element.node + seperator;
-      output += message.element.id + seperator;
-      output += message.element.class + seperator;
-      output += message.position.lineNumber + seperator;
-      output += message.position.columnNumber + seperator;
-      output += message.description + '\n';
-    });
-  });
-
-  return output;
-
 };
 
 export default (reports, options = defaultOptions) => {
@@ -84,11 +60,11 @@ class ReportsGenerator {
   }
 
   reportTxt(reports) {
-    return generateReport(reports, '|');
+    return GenerateReport(reports, '|');
   }
 
   reportCsv(reports) {
-    return generateReport(reports, ',');
+    return GenerateReport(reports, ',');
   }
 
   writeFile() {
